Added table and column comment support.
Also inlined colspecs code, because there wasn't much and it
facilitated separating literal processing into a function.
