--- pybigquery/colspecs.py
@@ -1,42 +1,0 @@
-# Copyright (c) 2021 The PyBigQuery Authors
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy of
-# this software and associated documentation files (the "Software"), to deal in
-# the Software without restriction, including without limitation the rights to
-# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-# the Software, and to permit persons to whom the Software is furnished to do so,
-# subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-"""
-Overrides for type-specific behavior
-"""
-
-import sqlalchemy.sql.sqltypes
-
-class BQString(sqlalchemy.sql.sqltypes.String):
-
-    def literal_processor(self, dialect):
-
-        def bqstring_process_literal(value):
-            if value:
-                value = repr(value.replace("%", "%%"))
-                if value[0] == '"':
-                    value = "'" + value[1:-1].replace("'", "\'") + "'"
-
-            return value
-
-        return bqstring_process_literal
-
-
-colspecs = {
-    sqlalchemy.sql.sqltypes.String: BQString,
-}
